podcastic/requirements.txt
requests==2.31.0
langchain==0.1.12
langgraph==0.0.26
python-dotenv==1.0.1
typer[all]==0.9.0
openai==1.12.0
pyyaml==6.0.1
pydub==0.25.1